LayoutTests/fast/forms/month-multiple-fields/month-multiple-fields-ax-aria-attributes.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<input id="test" type="month" value="2012-10">
<script>
description('This test checks aria-help attribute of fields in multiple fields week input UI.');

function focusedFieldValueDescription()
{
    var element = accessibilityController.focusedElement;
    return element.helpText + ', ' +  element.valueDescription + ', ' + element.minValue + ', ' + element.maxValue;
}

var testInput = document.getElementById('test');

if (!window.accessibilityController || !window.eventSender)
    debug('Please run inside DRT or WTR.');
else {
    debug('Non-empty value');
    testInput.focus();
    shouldBeEqualToString('focusedFieldValueDescription()', 'AXHelp: Month, AXValueDescription: October, 1, 12');
    eventSender.keyDown('\t');
    shouldBeEqualToString('focusedFieldValueDescription()', 'AXHelp: Year, AXValueDescription: 2012, 1, 275760');

    debug('Empty value');
    eventSender.keyDown('\b');
    shouldBeEqualToString('focusedFieldValueDescription()', 'AXHelp: Year, AXValueDescription: blank, 1, 275760');
    eventSender.keyDown('\t', ['shiftKey']);
    eventSender.keyDown('\b');
    shouldBeEqualToString('focusedFieldValueDescription()', 'AXHelp: Month, AXValueDescription: blank, 1, 12');

    debug('');
    testInput.remove();
}
</script>
<script src="../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
